correção da interface AuthController e sua implementação AuthControllerImpl
