+++ src/main/java/br/com/squadjoaquina/errorlogger/controller/impl/AuthControllerImpl.java
@@ -1,16 +1,19 @@
-package br.com.squadjoaquina.errorlogger.controller.impl;
+package br.com.squadjoaquina.errorlogger.controller;
 
-import br.com.squadjoaquina.errorlogger.controller.AuthController;
+import br.com.squadjoaquina.errorlogger.dto.EmailDTO;
 import br.com.squadjoaquina.errorlogger.dto.UserDTO;
 import br.com.squadjoaquina.errorlogger.security.JWTUtil;
+import br.com.squadjoaquina.errorlogger.service.AuthService;
 import br.com.squadjoaquina.errorlogger.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletResponse;
+import javax.validation.Valid;
 
 @RestController
 @RequestMapping(value = "/auth")
@@ -18,7 +21,8 @@ public class AuthControllerImpl implements AuthController {
 
     @Autowired
     private JWTUtil jwtUtil;
-
+    @Autowired
+    private AuthService authService;
 
     @RequestMapping(value = "/refresh_token", method = RequestMethod.POST)
     public ResponseEntity<Void> refreshToken(HttpServletResponse response) {
@@ -26,8 +30,13 @@ public class AuthControllerImpl implements AuthController {
         String token = jwtUtil.generateToken(user.getUsername());
         response.addHeader("Authorization", "Bearer " + token);
         response.addHeader("access-control-expose-headers", "Authorization");
-        return ResponseEntity.noContent()
-                             .build();
+        return ResponseEntity.noContent().build();
+    }
+
+    @RequestMapping(value = "/forgot", method = RequestMethod.POST)
+    public ResponseEntity<Void> forgot(@Valid @RequestBody EmailDTO emailDTO) {
+        authService.sendNewPassword(emailDTO.getEmail());
+        return ResponseEntity.noContent().build();
     }
